feat: ajustes em arquivo models
