+++ app/models/index.ts
@@ -1,0 +1,10 @@
+export const CorBanco = [
+  "#3B82F6", // blue
+  "#10B981", // emerald
+  "#8B5CF6", // violet
+  "#F59E0B", // amber
+  "#EF4444", // red
+  "#06B6D4", // cyan
+  "#84CC16", // lime
+  "#F97316", // orange
+];
